feat(v0.7): add the operator operations service
Adds internal/panelpg, the PostgreSQL adapter behind the panel's day-to-day
work: the dashboard and its metric definitions, customer search and profile,
subscription detail, catalogue and promotion administration, finance search
and CSV export, fulfillment and webhook diagnostics, blocklist adjudication,
anomaly rules and evaluation, gifts, personal offers, saved payment methods
and dunning review, the digital-goods shop, and previewed bulk actions.

Every mutation writes its audit event in the same transaction as the change,
and no risk surface can apply an adverse action on its own: a blocklist match
and an anomaly signal produce rows an operator reviews.

+++ database/queries/risk.sql
@@ -286,11 +286,11 @@ HAVING count(*) >= sqlc.arg(minimum_count);
 -- The distinction is why the rule's window only controls how often a signal is
 -- re-raised, and it is documented on the operator page as well.
 SELECT s.id AS subscription_id, s.user_id,
-       COALESCE((s.observed_state->>'usedTrafficBytes')::bigint, 0) AS used_bytes,
+       (COALESCE((s.observed_state->>'usedTrafficBytes')::bigint, 0))::bigint AS used_bytes,
        s.reconciled_at
 FROM subscriptions s
 WHERE s.status = 'active'
   AND s.observed_state ? 'usedTrafficBytes'
-  AND COALESCE((s.observed_state->>'usedTrafficBytes')::bigint, 0) >= sqlc.arg(threshold_bytes)
+  AND COALESCE((s.observed_state->>'usedTrafficBytes')::bigint, 0) >= sqlc.arg(threshold_bytes)::bigint
 ORDER BY used_bytes DESC
 LIMIT sqlc.arg(page_size);
